% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{gc_call}
\alias{gc_call}
\title{call a running geocoding container to geocode an address}
\usage{
gc_call(address)
}
\arguments{
\item{address}{a string; see vignette for best practices and examples}
}
\description{
This is used internally by \code{geocode()} and normally should not be called directly.
}
