% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{start_geocoder_container}
\alias{start_geocoder_container}
\title{start geocoding container}
\usage{
start_geocoder_container(
  image_name = getOption("offlinegeocoder.image", "ghcr.io/degauss-org/geocoder:v3.4.0")
)
}
\arguments{
\item{image_name}{Name of geocoding image; can be used to specify the version. Defaults to \code{getOption( "offlinegeocoder.image", "ghcr.io/degauss-org/geocoder:v3.4.0")}.}
}
\description{
Beginning with v2.2, the latest version of the geocoder image is available
on the project GitHub repository:
\url{https://github.com/degauss-org/geocoder/pkgs/container/geocoder}
}
