% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{geocode}
\usage{
geocode(addresses, ...)
}
\arguments{
\item{addresses}{a list or vector addresses}

\item{...}{additional arguments passed to \code{CB::mappp}; set options for
cache here}
}
\value{
list of geocoding results with one address per element; some
  addresses may return more than one geocoding result if there is a tie among
  the best matches
}
\description{
To prevent errors and optimize accuracy, please remove non alphanumeric
characters from address strings prior to geocoding. In general, address
cleaning is outside the scope of this package and should be completed prior
to geocoding.
}
\details{
By default, geocoding results will be cached in a local folder named
\code{geocoding_cache}.  See help for \code{CB::mappp} to change these
defaults.
}
