% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{start_geocoder_container}
\alias{start_geocoder_container}
\title{start geocoding container}
\usage{
start_geocoder_container(image_name = "degauss/geocoder")
}
\arguments{
\item{image_name}{name of geocoding image; can be used to specify the version
i.e. \code{degauss/geocoder_slim:2.4}; defaults to
\code{degauss/geocoder_slim:latest}}
}
\description{
start geocoding container
}
